refactor: reorganize admin and debug commands
