+++ commands/admin/admin_battleinspect.py
